fix: modeling inbox page
