+++ apps/frontend/src/lib/store/inbox.store.ts
@@ -31,7 +31,7 @@ const useInboxStore = create<InboxStoreType>((set) => ({
       inboxItems_ = data.response as InboxItem[]
     } catch (error) {
       const e = error as AxiosError
-      console.error("Error while fetching my::", e.message)
+      console.error("error fetching my: ", e)
     }
     set({ inboxItems: inboxItems_ })
     return inboxItems_
@@ -52,7 +52,7 @@ const useInboxStore = create<InboxStoreType>((set) => ({
       todayItems_ = data.response.items as TodayInboxItem[]
     } catch (error) {
       const e = error as AxiosError
-      console.error("Error while fetching my-today::", e.message)
+      console.error("error fetching my-today: ", e)
     }
     set({ todayInboxItems: todayItems_ })
     return todayItems_
@@ -73,7 +73,7 @@ const useInboxStore = create<InboxStoreType>((set) => ({
       overdueItems_ = data.response.items as OverdueInboxItem[]
     } catch (error) {
       const e = error as AxiosError
-      console.error("Error while fetching my-overdue::", e.message)
+      console.error("error fetching my-overdue: ", e)
     }
     set({ overdueInboxItems: overdueItems_ })
     return overdueItems_
@@ -105,14 +105,14 @@ const useInboxStore = create<InboxStoreType>((set) => ({
       return updatedItem
     } catch (error) {
       const e = error as AxiosError
-      console.error("Error while moving item to new date::", e.message)
+      console.error("error moving item to new date: ", e)
       return null
     }
   },
-  updateItem: (editedItem: InboxItem, id: string) => {
+  updateItem: async (session: string, editedItem: InboxItem, id: string) => {
     set((state) => ({
       inboxItems: state.inboxItems.map((item) =>
-        item.uuid === id
+        item._id === id
           ? {
               ...item,
               title: editedItem.title,
@@ -121,6 +121,16 @@ const useInboxStore = create<InboxStoreType>((set) => ({
           : item
       ),
     }))
+    try {
+      await axios.put(`${BACKEND_URL}/api/items/${id}`, editedItem, {
+        headers: {
+          Authorization: `Bearer ${session}`,
+        },
+      })
+    } catch (error) {
+      const e = error as AxiosError
+      console.error("error updating inbox item: ", e)
+    }
   },
   addItem: async (session: string, title: string, description: string) => {
     let res: InboxItemCreateResponse
@@ -144,7 +154,7 @@ const useInboxStore = create<InboxStoreType>((set) => ({
       return res.item
     } catch (error) {
       const e = error as AxiosError
-      console.error("error adding item to inbox:", e)
+      console.error("error adding item to inbox: ", e)
       return null
     }
   },
@@ -172,7 +182,7 @@ const useInboxStore = create<InboxStoreType>((set) => ({
       )
     } catch (error) {
       const e = error as AxiosError
-      console.error("error deleting inbox item:", e)
+      console.error("error deleting inbox item: ", e)
     }
   },
 }))
